Re-add kernel_exec+data_mov measurement.
Report both only kernel exec, and kernel exec + data movement.

+++ velfg/src/kernel_velfg_ndrange.hpp
@@ -10,7 +10,7 @@
 
 using namespace sycl;
 
-int velfg_ndrange(queue &q, const std::vector<float> &u, const std::vector<float> &v,
+double velfg_ndrange(queue &q, const std::vector<float> &u, const std::vector<float> &v,
                    const std::vector<float> &w, const std::vector<float> &dx1,
                    const std::vector<float> &dy1, const std::vector<float> &dzn,
                    const std::vector<float> &dzs, std::vector<float> &f, std::vector<float> &g,
@@ -454,7 +454,7 @@ int velfg_ndrange(queue &q, const std::vector<float> &u, const std::vector<float
 
   auto start = event_first.get_profiling_info<info::event_profiling::command_start>();
   auto end = event_last.get_profiling_info<info::event_profiling::command_end>();
-  int time_in_ms = static_cast<int>(static_cast<ulong>(end - start) / 1000000u);
+  double time_in_ms = static_cast<double>(end - start) / 1000000;
 
   return time_in_ms;
 }
