+++ velfg/src/kernel_velfg_swi.hpp
@@ -10,7 +10,7 @@
 
 using namespace sycl;
 
-int velfg_swi(queue &q, const std::vector<float> &u, const std::vector<float> &v,
+double velfg_swi(queue &q, const std::vector<float> &u, const std::vector<float> &v,
                const std::vector<float> &w, const std::vector<float> &dx1,
                const std::vector<float> &dy1, const std::vector<float> &dzn,
                const std::vector<float> &dzs, std::vector<float> &f, std::vector<float> &g,
@@ -438,7 +438,7 @@ int velfg_swi(queue &q, const std::vector<float> &u, const std::vector<float> &v
 
   auto start = event_first.get_profiling_info<info::event_profiling::command_start>();
   auto end = event_last.get_profiling_info<info::event_profiling::command_end>();
-  int time_in_ms = static_cast<int>(static_cast<ulong>(end - start) / 1000000u);
+  double time_in_ms = static_cast<double>(end - start) / 1000000;
 
   return time_in_ms;
 }
